Renamed Parser namespace to BodyParser and also moved the FormParserFactory: https://github.com/reactphp/http/pull/41#discussion_r69938047
